graph-service/requirements.txt
fastapi
uvicorn
neo4j
networkx
scipy
numpy
python-dotenv
prometheus_fastapi_instrumentator
psycopg2-binary
